backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
python-multipart==0.0.6
google-api-python-client==2.108.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.7
supabase==2.1.0
email-validator==2.3.0
schedule==1.2.0
